+++ src/app/budget/animations/route-animations.ts
@@ -1,0 +1,173 @@
+import {
+  transition,
+  style,
+  animate,
+  trigger,
+  group,
+  query
+} from '@angular/animations';
+
+export const routerTransitionTrigger = trigger('routerTransition', [
+  transition('budget-dashboard => budget-lines', [
+    /* order */
+    /* 1 */ query(
+      ':enter, :leave',
+      style({ position: 'fixed', width: '100%' }),
+      { optional: true }
+    ),
+    /* 2 */ group([
+      // block executes in parallel
+      query(
+        ':enter',
+        [
+          style({ transform: 'translateX(100%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(0%)' }))
+        ],
+        { optional: true }
+      ),
+      query(
+        ':leave',
+        [
+          style({ transform: 'translateX(0%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(-100%)' }))
+        ],
+        { optional: true }
+      )
+    ])
+  ]),
+  transition('budget-lines => budget-dashboard', [
+    /* order */
+    /* 1 */ query(
+      ':enter, :leave',
+      style({ position: 'fixed', width: '100%' }),
+      { optional: true }
+    ),
+    /* 2 */ group([
+      // block executes in parallel
+      query(
+        ':enter',
+        [
+          style({ transform: 'translateX(-100%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(0%)' }))
+        ],
+        { optional: true }
+      ),
+      query(
+        ':leave',
+        [
+          style({ transform: 'translateX(0%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(100%)' }))
+        ],
+        { optional: true }
+      )
+    ])
+  ]),
+  transition('budget-lines => budget-alerts', [
+    /* order */
+    /* 1 */ query(
+      ':enter, :leave',
+      style({ position: 'fixed', width: '100%' }),
+      { optional: true }
+    ),
+    /* 2 */ group([
+      // block executes in parallel
+      query(
+        ':enter',
+        [
+          style({ transform: 'translateX(100%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(0%)' }))
+        ],
+        { optional: true }
+      ),
+      query(
+        ':leave',
+        [
+          style({ transform: 'translateX(0%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(-100%)' }))
+        ],
+        { optional: true }
+      )
+    ])
+  ]),
+  transition('budget-alerts => budget-lines', [
+    /* order */
+    /* 1 */ query(
+      ':enter, :leave',
+      style({ position: 'fixed', width: '100%' }),
+      { optional: true }
+    ),
+    /* 2 */ group([
+      // block executes in parallel
+      query(
+        ':enter',
+        [
+          style({ transform: 'translateX(-100%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(0%)' }))
+        ],
+        { optional: true }
+      ),
+      query(
+        ':leave',
+        [
+          style({ transform: 'translateX(0%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(100%)' }))
+        ],
+        { optional: true }
+      )
+    ])
+  ]),
+  transition('budget-alerts => budget-dashboard', [
+    /* order */
+    /* 1 */ query(
+      ':enter, :leave',
+      style({ position: 'fixed', width: '100%' }),
+      { optional: true }
+    ),
+    /* 2 */ group([
+      // block executes in parallel
+      query(
+        ':enter',
+        [
+          style({ transform: 'translateX(100%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(0%)' }))
+        ],
+        { optional: true }
+      ),
+      query(
+        ':leave',
+        [
+          style({ transform: 'translateX(0%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(-100%)' }))
+        ],
+        { optional: true }
+      )
+    ])
+  ]),
+  transition('budget-dashboard => budget-alerts', [
+    /* order */
+    /* 1 */ query(
+      ':enter, :leave',
+      style({ position: 'fixed', width: '100%' }),
+      { optional: true }
+    ),
+    /* 2 */ group([
+      // block executes in parallel
+      query(
+        ':enter',
+        [
+          style({ transform: 'translateX(-100%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(0%)' }))
+        ],
+        { optional: true }
+      ),
+      query(
+        ':leave',
+        [
+          style({ transform: 'translateX(0%)' }),
+          animate('0.5s ease-in-out', style({ transform: 'translateX(100%)' }))
+        ],
+        { optional: true }
+      )
+    ])
+  ])
+]);